Move tcp_tracer to a separate directory
