+++ src/core/ext/transport/chttp2/transport/context_list_entry.h
@@ -28,7 +28,7 @@
 #include <utility>
 #include <vector>
 
-#include "src/core/lib/channel/tcp_tracer.h"
+#include "src/core/lib/channel/tcp_tracer/tcp_tracer.h"
 
 namespace grpc_core {
