+++ src/cpp/ext/filters/census/open_census_call_tracer.h
@@ -40,7 +40,7 @@
 
 #include "src/core/lib/channel/call_tracer.h"
 #include "src/core/lib/channel/context.h"
-#include "src/core/lib/channel/tcp_tracer.h"
+#include "src/core/lib/channel/tcp_tracer/tcp_tracer.h"
 #include "src/core/lib/gprpp/sync.h"
 #include "src/core/lib/iomgr/error.h"
 #include "src/core/lib/resource_quota/arena.h"
