+++ src/cpp/ext/otel/otel_call_tracer.h
@@ -34,7 +34,7 @@
 #include <grpc/support/time.h>
 
 #include "src/core/lib/channel/call_tracer.h"
-#include "src/core/lib/channel/tcp_tracer.h"
+#include "src/core/lib/channel/tcp_tracer/tcp_tracer.h"
 #include "src/core/lib/gprpp/sync.h"
 #include "src/core/lib/iomgr/error.h"
 #include "src/core/lib/resource_quota/arena.h"
